Initial commit. Details follow.
This is the initial commit, which is being made to easily follow code progression. Though this code will (should) compile and run, it does not currently perform the intended actions. Everything is being added, because I don't have enough experience with Visual Studio to know what not to include. I'm sure I'll figure it out some day.